+++ apps/designer-standalone/src/app/Designer/designer.tsx
@@ -10,6 +10,7 @@ import {
   StandardGatewayService,
   StandardRunService,
 } from '@microsoft/designer-client-services-logic-apps';
+import type { ContentType } from '@microsoft/designer-client-services-logic-apps';
 import { DesignerProvider, BJSWorkflowProvider, Designer } from '@microsoft/logic-apps-designer';
 import { ResourceIdentityType } from '@microsoft/utils-logic-apps';
 import { useEffect } from 'react';
@@ -75,12 +76,23 @@ const runService = new StandardRunService({
 
 const workflowService = { getCallbackUrl: () => Promise.resolve({ method: 'POST', value: 'Dummy url' }) };
 
+const hostService = { fetchAndDisplayContent: (title: string, url: string, type: ContentType) => console.log(title, url, type) };
+
 export const DesignerWrapper = () => {
   const { workflowDefinition, readOnly, monitoringView, darkMode, consumption, connections, runInstance } = useSelector(
     (state: RootState) => state.workflowLoader
   );
   const designerProviderProps = {
-    services: { connectionService, operationManifestService, searchService, oAuthService, gatewayService, workflowService, runService },
+    services: {
+      connectionService,
+      operationManifestService,
+      searchService,
+      oAuthService,
+      gatewayService,
+      workflowService,
+      hostService,
+      runService,
+    },
     readOnly,
     isMonitoringView: monitoringView,
     isDarkMode: darkMode,