+++ libs/designer/src/lib/core/state/designerOptions/designerOptionsInterfaces.ts
@@ -7,6 +7,7 @@ import type {
   ISearchService,
   IOAuthService,
   IWorkflowService,
+  IHostService,
   IApiManagementService,
   IRunService,
 } from '@microsoft/designer-client-services-logic-apps';
@@ -28,6 +29,7 @@ export interface ServiceOptions {
   loggerService?: ILoggerService;
   oAuthService: IOAuthService;
   workflowService: IWorkflowService;
+  hostService?: IHostService;
   apimService?: IApiManagementService;
   runService?: IRunService;
 }