+++ libs/services/designer-client-services/src/index.ts
@@ -13,5 +13,6 @@ export * from './lib/logging/logEntry';
 export * from './lib/base';
 export * from './lib/standard';
 export * from './lib/consumption';
+export * from './lib/host';
 export * from './lib/apimanagement';
 export * from './lib/run';